Authorization from Frontend

+++ backend3/books.js
@@ -1,3 +1,6 @@
+const { verify } = require("jsonwebtoken");
+const jwt = require("jsonwebtoken")
+
 function booksSQLite(app, db) {
     //creating a table called "books"
     db.run(`
@@ -9,6 +12,7 @@ function booksSQLite(app, db) {
   `);
     // Selecting All
     app.get('/api/books', (req, res) => {
+        if (!verifyToken(req, res)) return
         const sql = "SELECT * FROM books"
         db.all(sql, (error, result) => {
             if (error) return res.json(error.message)
@@ -17,6 +21,7 @@ function booksSQLite(app, db) {
     })
     // Selecting by id
     app.get('/api/books/:id', (req, res) => {
+        if (!verifyToken(req, res)) return
         const { id } = req.params
         const sql = " SELECT * FROM books WHERE id=?"
         db.get(sql, [id], (error, result) => {
@@ -26,6 +31,7 @@ function booksSQLite(app, db) {
     })
     // Adding / inserting book into database
     app.post('/api/books', (req, res) => {
+        if (!verifyToken(req, res)) return
         const { name, price } = req.body
         const sql = "INSERT INTO books (name, price) VALUES (?,?)"
         db.run(sql, [name, price], (error) => {
@@ -35,6 +41,7 @@ function booksSQLite(app, db) {
     })
     // updating by id 
     app.post('/api/books/:id', (req, res) => {
+        if (!verifyToken(req, res)) return
         const { id } = req.params
         const { name, price } = req.body
         const sql = " UPDATE books  set name=?, price=? WHERE id=?"
@@ -45,6 +52,7 @@ function booksSQLite(app, db) {
     })
     // Delete all
     app.delete('/api/books', (req, res) => {
+        if (!verifyToken(req, res)) return
         const sql = "DELETE FROM books"
         db.run(sql, (error) => {
             if (error) return res.json(error.message)
@@ -53,6 +61,7 @@ function booksSQLite(app, db) {
     })
     // Delete only one
     app.delete('/api/books/:id', (req, res) => {
+        if (!verifyToken(req, res)) return
         const { id } = req.params
         const sql = "DELETE FROM books WHERE id=?"
         db.run(sql, [id], (error) => {
@@ -60,6 +69,21 @@ function booksSQLite(app, db) {
             return res.json(" id +  books deleted")
         })
     })
+    // Token verification 
+    function verifyToken(req, res) {
+        const token = req?.headers?.authorization?.split(" ")[1];
+        try {
+            return jwt.verify(token, "RANDOM-TOKEN");
+        } catch (error) {
+            if (error instanceof jwt.TokenExpiredError) {
+                res.status(401).json({ error: "Token has expired" });
+            } else {
+                res.status(500).json({ error: "Token verification failed" });
+            }
+            return false
+        }
+    }
+
 }
 
 module.exports = { booksSQLite }
