+++ backend3/students.js
@@ -1,3 +1,4 @@
+const jwt = require("jsonwebtoken")
 function studentsSQLiteAPI(app, db) {
     db.run(`
   CREATE TABLE IF NOT EXISTS students (
@@ -27,6 +28,7 @@ function studentsSQLiteAPI(app, db) {
     })
     // adding or insering students into the talbe
     app.post('/students', (req, res) => {
+        if (!verifyToken(req, res)) return
         const { name, phone } = req.body
         const sql = "INSERT INTO students (name, phone) VALUES (?,?)"
         db.run(sql, [name, phone], (error) => {
@@ -45,6 +47,7 @@ function studentsSQLiteAPI(app, db) {
     })*/
 
     app.delete('/students/:id', (req, res) => {
+        if (!verifyToken(req, res)) return
         const { id } = req.params
         const sql = "DELETE FROM students where id=?"
         db.run(sql, [id], (error) => {
@@ -55,6 +58,7 @@ function studentsSQLiteAPI(app, db) {
 
     // Deleting all the students
     app.delete('/students', (req, res) => {
+        if (!verifyToken(req, res)) return
         const { id } = req.params
         const sql = "DELETE FROM students"
         db.run(sql, (error) => {
@@ -73,14 +77,30 @@ function studentsSQLiteAPI(app, db) {
         })
     })*/
     app.post('/students/:id', (req, res) => {
+        if (!verifyToken(req, res)) return
         const { id } = req.params
         const { name, phone } = req.body
         const sql = "UPDATE students set name=?, phone=? where id=?"
         db.run(sql, [name, phone, id], (error) => {
             if (error) return res.status(500).json(error.message)
-            return res.status(200).json("Updated ssss ")
+            return res.status(200).json(" Student updated Successfully ")
         })
 
     })
+    // Token verification 
+    function verifyToken(req, res) {
+        const token = req?.headers?.authorization?.split(" ")[1];
+        try {
+            return jwt.verify(token, "RANDOM-TOKEN");
+        } catch (error) {
+            if (error instanceof jwt.TokenExpiredError) {
+                res.status(401).json({ error: "Token has expired" });
+            } else {
+                res.status(500).json({ error: "Token verification failed" });
+            }
+            return false
+        }
+    }
+
 }
 module.exports = { studentsSQLiteAPI }
